Update AbstractContentViewHelper.php
Its possible to give in an vh (eg. v:content.get) the wildcard '%' to parameter "column". With this now it starts to be possible to to get *all* contentelements on the page and so f.e the option "sectionIndexOnly" starts to make sense. Otherwise the column allways will be set to zero and no CE is rendered. Or you can give a column number but all other CE's which might be in other grids etc. are not included. For all other case the Operator "=" still is used.